Update jobs.py
While creating a job which itself has a job as a node i am unable to pass notebook params to the job for that JobRun. After adding this value I can see i am able to pass the notebook_params to the RUN node.  The issue is that the main job's global params are not read by the notebook inside the sub-run

Signed-off-by: Chaitanya Sharma <52489647+chaitti@users.noreply.github.com>